[stdlib] Add better support for BinaryInteger.words
Removes BinaryInteger's _word(at:) requirement and its
countOfRepresentedWords method, making the words property the sole way
to access an integer's words.
This is a better fix for https://bugs.swift.org/browse/SR-5275.

Existing code that needs word-level access to integers now uses
the words property. However, the Words associated type cannot be
explicitly constrained to be a collection yet, so code in generic
context needs some sort of workaround to be able to work with its
values. (This affects default conversion initializers on
FixedWidthIntegers.)

The new BinaryInteger._wordsIterator() requirement provides access to
an iterator for the words of an integer. It has a default
implementation in terms of the words property whenever Word satisfies
the documented constraints. This is intended as temporary and should
be removed when it becomes possible to constrain Words explicitly.