[stdlib] Create a new generic DefaultWords collection; use it in all standard fixed-width integer types.
This unifies the word types of standard integers on 32- and 64-bit platforms.
(So that, for example, Int64.Words is the same type on all platforms, even
though it contains one or two words depending on the word size.)

DefaultWords works for all fixed-width integer types, so set it as the default
value for FixedWidthInteger.Words.